fix: 기존의 인-메모리 DB 형태였던 구현체 ReservationRepository를 추상화한 인터페이스 객체 선언, 기존과 이번 JDBC 레포지토리 객체를 해당 인터페이스의 구현체로 등록
