+++ src/main/java/roomescape/repository/ReservationRepository.java
@@ -3,39 +3,15 @@ package roomescape.repository;
 import roomescape.entity.Reservation;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicLong;
 
-public class ReservationRepository {
+public interface ReservationRepository {
 
-    private Map<Long, Reservation> reservations = new ConcurrentHashMap<>();
-    private AtomicLong index = new AtomicLong(0);
+    Reservation save(Reservation reservation);
 
-    public Reservation save(Reservation reservation) {
+    List<Reservation> findAll();
 
-        Reservation createdReservation = Reservation.toEntityWithId(index.incrementAndGet(), reservation);
-        this.reservations.put(createdReservation.getId(), createdReservation);
+    void delete(Reservation reservation);
 
-        return createdReservation;
-    }
-
-    public List<Reservation> findAll() {
-        return this.reservations.values().stream().toList();
-    }
-
-    public void delete(Reservation reservation) {
-        if (reservation == null || reservation.getId() == null) {
-            throw new IllegalArgumentException();
-        }
-        if (!reservations.containsKey(reservation.getId())) {
-            return;
-        }
-        this.reservations.remove(reservation.getId());
-    }
-
-    public Optional<Reservation> findById(Long reservationId) {
-        return Optional.ofNullable(reservations.get(reservationId));
-    }
+    Optional<Reservation> findById(Long reservationId);
 }
