+++ src/main/java/roomescape/service/ReservationServiceImpl.java
@@ -3,15 +3,15 @@ package roomescape.service;
 import roomescape.dto.request.ReservationRequest;
 import roomescape.entity.Reservation;
 import roomescape.exception.ReservationNotFoundException;
-import roomescape.repository.ReservationRepository;
+import roomescape.repository.ReservationRepositoryV1;
 
 import java.util.List;
 
 public class ReservationServiceImpl implements ReservationService {
 
-    private final ReservationRepository reservationRepository;
+    private final ReservationRepositoryV1 reservationRepository;
 
-    public ReservationServiceImpl(ReservationRepository reservationRepository) {
+    public ReservationServiceImpl(ReservationRepositoryV1 reservationRepository) {
         this.reservationRepository = reservationRepository;
     }
